Move path-utils to TypeChain core package
