--- packages/typechain/src/utils/files.ts
@@ -1,20 +1,0 @@
-import { isAbsolute, join, parse } from 'path'
-
-export function getFilename(path: string) {
-  return parse(path).name
-}
-
-export function getFileExtension(path: string) {
-  return parse(path).ext
-}
-
-export function ensureAbsPath(path: string): string {
-  if (isAbsolute(path)) {
-    return path
-  }
-  return join(process.cwd(), path)
-}
-
-export function normalizeSlashes(path: string) {
-  return process.platform === 'win32' ? path.replace(/\\/g, '/') : path
-}
