+++ packages/typechain/src/utils/files/getFileExtension.ts
@@ -1,0 +1,6 @@
+import { parse } from 'path';
+
+
+export function getFileExtension(path: string) {
+  return parse(path).ext;
+}
