+++ packages/typechain/src/utils/files/getFilename.ts
@@ -1,0 +1,6 @@
+import { parse } from 'path';
+
+
+export function getFilename(path: string) {
+  return parse(path).name;
+}
