+++ packages/typechain/src/utils/files/lowestCommonPath.ts
@@ -1,0 +1,5 @@
+export function lowestCommonPath(paths: string[]) {
+  const pathParts = paths.map((path) => path.split(/[\\/]/))
+  const commonParts = pathParts[0].filter((part, index) => pathParts.every((parts) => parts[index] === part))
+  return commonParts.join('/')
+}
