+++ packages/typechain/src/utils/files/normalizeSlashes.ts
@@ -1,0 +1,4 @@
+
+export function normalizeSlashes(path: string) {
+  return process.platform === 'win32' ? path.replace(/\\/g, '/') : path;
+}
